chore: move service request page inside directory
